Key Points  * Recorder threads are similar to the model / setup of the receiver threads    * Created at initialization - no spawning and despawning    * Each thread is responsible for the assigned antennas, written to 1 file  * Receivers post to dispatcher thread  * Main dispatcher thread (core 0) will assign messages to the applicable recorder threads  * With 1 recorder thread the file will be identical to the legacy file
Changes
 * The stored array will only have the antennas in the file as max dimension (if the file only has two antennas that will be t the ant dimension of the dataset)
 * Added the following HD5f attributes
   ANT_OFFSET = Starting antenna number
   ANT_NUM    = Total number of antennas represented in the file
   ANT_TOTAL  = Total number of antennas in the experiment
 * Added the following configuration value
   task_thread = Number of recorder threads (probably should change this name)
 * Appends first and last antenna number to the end of the file name
 * >1 recorder thread will create files that have datasets with reduced antenna dimensions (0 = start antenna / offset)
