+++ CC/Sounder/CMakeLists.txt
@@ -126,6 +126,8 @@ set(SOUNDER_SOURCES
     Radio.cc
     receiver.cc
     recorder.cc
+    recorder_worker.cc
+    recorder_thread.cc
     BaseRadioSet.cc
     BaseRadioSet-calibrate.cc
     comms-lib.cc
